Extract BindingTestCase from RecordFactory
